Add editor and experimental settings tabs with corresponding components
- Introduced new settings tabs for "Editor" and "Experimental" in the settings interface.
- Added `EditorSettingsPane` and `ExperimentalSettingsPane` components to manage respective settings.
- Updated `MainContent` to include the new tabs and their content.
- Adjusted sidebar settings to reflect the new tab structure.
- Enhanced localization for new settings sections in multiple languages.
- Removed deprecated settings related to editor configurations from the GeneralSettingsPane.

These changes improve the organization of settings, allowing users to access editor and experimental features more intuitively.
