+++ src/components/settings/settingsSearch.ts
@@ -41,9 +41,9 @@ const SETTINGS_SEARCH_DEFS: readonly SettingsSearchDef[] = [
 	{ id: "appearance-custom-themes", tab: "appearance" },
 	{ id: "appearance-interface-font", tab: "appearance" },
 	{ id: "appearance-monospace-font", tab: "appearance" },
-	{ id: "appearance-editor-font", tab: "appearance" },
+	{ id: "appearance-editor-font", tab: "editor" },
 	{ id: "appearance-ui-font-size", tab: "appearance" },
-	{ id: "appearance-editor-font-size", tab: "appearance" },
+	{ id: "appearance-editor-font-size", tab: "editor" },
 	{ id: "shortcuts-customize", tab: "shortcuts" },
 	{ id: "shortcuts-search", tab: "shortcuts" },
 	{ id: "shortcuts-workspace", tab: "shortcuts" },
@@ -89,23 +89,23 @@ const SETTINGS_SEARCH_DEFS: readonly SettingsSearchDef[] = [
 	{ id: "git-include-templates", tab: "git" },
 	{ id: "git-include-attachments", tab: "git" },
 	{ id: "git-include-non-markdown", tab: "git" },
-	{ id: "general-editor-table-of-contents", tab: "general" },
+	{ id: "general-editor-table-of-contents", tab: "editor" },
 	{ id: "space-search-index-people-tags", tab: "space" },
-	{ id: "general-editor-frontmatter", tab: "general" },
-	{ id: "general-editor-heading-prefixes", tab: "general" },
-	{ id: "general-editor-colorful-headings", tab: "general" },
-	{ id: "appearance-editor-presentation-beautiful-tags", tab: "appearance" },
-	{ id: "appearance-editor-presentation-width", tab: "appearance" },
-	{ id: "general-editor-collapsible-headings", tab: "general" },
-	{ id: "general-editor-collapsible-lists", tab: "general" },
-	{ id: "general-editor-spell-check", tab: "general" },
-	{ id: "general-editor-external-link-previews", tab: "general" },
-	{ id: "general-editor-vim-mode", tab: "general" },
+	{ id: "general-editor-frontmatter", tab: "editor" },
+	{ id: "general-editor-heading-prefixes", tab: "editor" },
+	{ id: "general-editor-colorful-headings", tab: "editor" },
+	{ id: "appearance-editor-presentation-beautiful-tags", tab: "editor" },
+	{ id: "appearance-editor-presentation-width", tab: "editor" },
+	{ id: "general-editor-collapsible-headings", tab: "editor" },
+	{ id: "general-editor-collapsible-lists", tab: "editor" },
+	{ id: "general-editor-spell-check", tab: "editor" },
+	{ id: "general-editor-external-link-previews", tab: "experimental" },
+	{ id: "general-editor-vim-mode", tab: "experimental" },
 	{ id: "ai-assistant-behavior-tools", tab: "ai" },
-	{ id: "appearance-layout-folio-mode", tab: "appearance" },
-	{ id: "appearance-layout-classic-all-notes", tab: "appearance" },
+	{ id: "appearance-layout-folio-mode", tab: "experimental" },
+	{ id: "appearance-layout-classic-all-notes", tab: "experimental" },
 	{ id: "general-file-tree-folder-counts", tab: "general" },
-	{ id: "general-file-tree-non-markdown-files", tab: "general" },
+	{ id: "general-file-tree-non-markdown-files", tab: "experimental" },
 	{ id: "general-file-tree-sort", tab: "general" },
 	{ id: "appearance-database-column-color", tab: "appearance" },
 	{ id: "about-app", tab: "about" },
