+++ src/i18n/locales/en/settings.general.json
@@ -3,6 +3,7 @@
 		"tab": "General",
 		"panelTitle": "General",
 		"groupApplication": "Application",
+		"groupLabs": "Labs",
 		"searchPlaceholder": "Search settings",
 		"searchAriaLabel": "Search settings",
 		"clearSearchAriaLabel": "Clear settings search"
@@ -73,6 +74,10 @@
 			"description": "Closing the last window hides it while Glyph remains open. Use Quit Glyph to quit.",
 			"ariaLabel": "Keep Glyph running after closing the last window"
 		}
+	},
+	"experimental": {
+		"sectionTitle": "Experimental",
+		"sectionDescription": "Experimental features that are not fully fleshed out. A later version of Glyph may change or remove them."
 	},
 	"editor": {
 		"sectionTitle": "Editor",
