+++ src/i18n/locales/ja/settings.general.json
@@ -3,6 +3,7 @@
 		"tab": "一般",
 		"panelTitle": "一般",
 		"groupApplication": "アプリケーション",
+		"groupLabs": "ラボ",
 		"searchPlaceholder": "設定を検索",
 		"searchAriaLabel": "設定を検索",
 		"clearSearchAriaLabel": "設定の検索をクリア"
@@ -73,6 +74,10 @@
 			"description": "最後のウインドウを閉じると非表示になりますが、Glyph は開いたままです。終了するには「Glyph を終了」を使用します。",
 			"ariaLabel": "最後のウインドウを閉じても Glyph を実行し続ける"
 		}
+	},
+	"experimental": {
+		"sectionTitle": "実験的",
+		"sectionDescription": "まだ作り込みが足りない実験的な機能です。今後の Glyph のバージョンで変更または削除されることがあります。"
 	},
 	"editor": {
 		"sectionTitle": "エディタ",
